Render new mbed_config cmake template with information from target and config
