+++ mbed_build/_internal/cmake_file.py
@@ -3,17 +3,21 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 """Module in charge of CMake file generation."""
+import datetime
 import pathlib
 from typing import Iterable
 
 import jinja2
-from mbed_targets import get_target_by_name
+from mbed_targets import get_target_by_name, Target
+
+from mbed_build._internal.config.config import Config
+from mbed_build._internal.config.assemble_build_config import assemble_config
 
 TEMPLATES_DIRECTORY = pathlib.Path("_internal", "templates")
-TEMPLATE_NAME = "CMakeLists.tmpl"
+TEMPLATE_NAME = "mbed_config.tmpl"
 
 
-def generate_cmakelists_file(mbed_target: str, program_path: str, toolchain_name: str) -> str:
+def generate_mbed_config_cmake_file(mbed_target: str, program_path: str, toolchain_name: str) -> str:
     """Generate the top-level CMakeLists.txt file containing the correct definitions for a build.
 
     Args:
@@ -25,34 +29,51 @@ def generate_cmakelists_file(mbed_target: str, program_path: str, toolchain_name
         A string of rendered contents for the file.
     """
     target_build_attributes = get_target_by_name(mbed_target, program_path)
-    return _render_cmakelists_template(
-        target_build_attributes.labels,
-        target_build_attributes.features,
-        target_build_attributes.components,
-        toolchain_name,
-    )
+    config = assemble_config(mbed_target, pathlib.Path(program_path))
+    return _render_mbed_config_cmake_template(target_build_attributes, config, toolchain_name, mbed_target,)
 
 
-def _render_cmakelists_template(
-    target_labels: Iterable[str], feature_labels: Iterable[str], component_labels: Iterable[str], toolchain_name: str,
+def _render_mbed_config_cmake_template(
+    target_build_attributes: Target, config: Config, toolchain_name: str, target_name: str
 ) -> str:
-    """Loads the CMakeLists.txt file template and renders it with the correct details.
+    """Renders the mbed_config template with the relevant information.
 
     Args:
-        target_labels: target-specific magic mbed-os directory names that need to be included in the build
-        feature_labels: target-specific magic mbed-os feature directory names that need to be included in the build
-        component_labels: target-specific magic mbed-os component directory names that need to be included in the build
-        toolchain_name: the toolchain-specific magic mbed-os directory name that need to be included in the build
+        target_build_attributes: Target config object.
+        config: Config object holding information parsed from the mbed config system.
+        toolchain_name: Name of the toolchain being used.
+        target_name: Name of the target.
 
     Returns:
         The contents of the rendered CMake file.
     """
     env = jinja2.Environment(loader=jinja2.PackageLoader("mbed_build", str(TEMPLATES_DIRECTORY)),)
     template = env.get_template(TEMPLATE_NAME)
+    options = list(config.options.values())
+    macros = list(config.macros.values())
+
     context = {
-        "target_labels": target_labels,
-        "feature_labels": feature_labels,
-        "component_labels": component_labels,
+        "labels": target_build_attributes.labels,
+        "features": target_build_attributes.features,
+        "components": target_build_attributes.components,
+        "device_has": target_build_attributes.device_has,
+        "target_macros": target_build_attributes.macros,
+        "supported_form_factors": target_build_attributes.supported_form_factors,
+        "timestamp": datetime.datetime.now().timestamp(),
+        "core": target_build_attributes.core,
+        "target_name": target_name,
         "toolchain_name": toolchain_name,
+        "options": sorted(options, key=lambda option: option.macro_name),
+        "macros": sorted(macros, key=lambda macro: macro.name),
+        "max_name_length": max(_max_attribute_length(options, "macro_name"), _max_attribute_length(macros, "name")),
+        "max_value_length": max(_max_attribute_length(options, "value"), _max_attribute_length(macros, "value")),
     }
     return template.render(context)
+
+
+def _max_attribute_length(objects: Iterable[object], attribute: str) -> int:
+    attrs = (getattr(o, attribute) for o in objects)
+    try:
+        return max(len(str(attr)) for attr in attrs if attr is not None)
+    except ValueError:  # no attrs found
+        return 0
